+++ CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ensembles notebook ([#218](https://github.com/tinkoff-ai/etna-ts/pull/218))
 - Function plot_backtest_interactive ([#225](https://github.com/tinkoff-ai/etna-ts/pull/225))
 - Confidence intervals in Pipeline ([#221](https://github.com/tinkoff-ai/etna-ts/pull/221)) 
+- TreeFeatureSelectionTransform ([229](https://github.com/tinkoff-ai/etna-ts/pull/229))
 
 ### Changed
 - Delete offset from WindowStatisticsTransform ([#111](https://github.com/tinkoff-ai/etna-ts/pull/111))